Adjust tests for RPMPackage rename and changes
