+++ tests/models/test_advisory.py
@@ -21,12 +21,8 @@ from unittest import TestCase
 
 from notus.scanner.models.advisory import (
     Advisory,
-    OperatingSystemAdvisories,
-    PackageAdvisories,
-    PackageAdvisory,
     Severity,
 )
-from notus.scanner.models.package import Architecture, Package
 
 
 class AdvisoryTestCase(TestCase):
@@ -139,387 +135,3 @@ class AdvisoryTestCase(TestCase):
 
         with self.assertRaises(FrozenInstanceError):
             advisory.last_modification = datetime.now()
-
-
-class PackageAdvisoryTestCase(TestCase):
-    def test_constructor(self):
-        package = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-
-        package_advisory = PackageAdvisory(package=package, advisory=advisory)
-
-        self.assertEqual(package_advisory.package, package)
-        self.assertEqual(package_advisory.advisory, advisory)
-
-    def test_immutability(self):
-        package1 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        package2 = Package(
-            'bar', '1.2.3', '3', Architecture.AARCH64, 'bar-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory1 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-        advisory2 = Advisory(
-            oid="1.2.3.4.6",
-            title="Bar Foo",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="4321",
-            advisory_xref="http://bar/4321",
-            severity=severity,
-        )
-
-        package_advisory = PackageAdvisory(package=package1, advisory=advisory1)
-
-        with self.assertRaises(FrozenInstanceError):
-            package_advisory.package = package2
-
-        with self.assertRaises(FrozenInstanceError):
-            package_advisory.advisory = advisory2
-
-        self.assertEqual(package_advisory.package, package1)
-        self.assertEqual(package_advisory.advisory, advisory1)
-
-    def test_equal(self):
-        package1 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        package2 = Package(
-            'bar', '1.2.3', '3', Architecture.AARCH64, 'bar-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory1 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-        advisory2 = Advisory(
-            oid="1.2.3.4.6",
-            title="Bar Foo",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="4321",
-            advisory_xref="http://bar/4321",
-            severity=severity,
-        )
-
-        package_advisory1 = PackageAdvisory(
-            package=package1, advisory=advisory1
-        )
-        package_advisory2 = PackageAdvisory(
-            package=package2, advisory=advisory2
-        )
-        package_advisory3 = PackageAdvisory(
-            package=package1, advisory=advisory1
-        )
-        package_advisory4 = PackageAdvisory(
-            package=package1, advisory=advisory2
-        )
-
-        self.assertEqual(package_advisory1, package_advisory1)
-        self.assertEqual(package_advisory1, package_advisory3)
-        self.assertNotEqual(package_advisory2, package_advisory3)
-        self.assertNotEqual(package_advisory1, package_advisory2)
-        self.assertNotEqual(package_advisory1, package_advisory4)
-        self.assertNotEqual(package_advisory2, package_advisory4)
-
-    def test_hash(self):
-        package1 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        package2 = Package(
-            'bar', '1.2.3', '3', Architecture.AARCH64, 'bar-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory1 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-        advisory2 = Advisory(
-            oid="1.2.3.4.6",
-            title="Bar Foo",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="4321",
-            advisory_xref="http://bar/4321",
-            severity=severity,
-        )
-
-        package_advisory1 = PackageAdvisory(
-            package=package1, advisory=advisory1
-        )
-        package_advisory2 = PackageAdvisory(
-            package=package2, advisory=advisory2
-        )
-        package_advisory3 = PackageAdvisory(
-            package=package1, advisory=advisory1
-        )
-        package_advisory4 = PackageAdvisory(
-            package=package1, advisory=advisory2
-        )
-
-        self.assertEqual(hash(package_advisory1), hash(package_advisory1))
-        self.assertEqual(hash(package_advisory1), hash(package_advisory3))
-        self.assertNotEqual(hash(package_advisory1), hash(package_advisory2))
-        self.assertNotEqual(hash(package_advisory3), hash(package_advisory2))
-        self.assertNotEqual(hash(package_advisory1), hash(package_advisory4))
-        self.assertNotEqual(hash(package_advisory2), hash(package_advisory4))
-
-
-class PackageAdvisoriesTestCase(TestCase):
-    def test_constructor(self):
-        package_advisories = PackageAdvisories()
-
-        self.assertIsInstance(package_advisories.advisories, dict)
-        self.assertEqual(len(package_advisories), 0)
-
-    def test_add_advisory_for_package(self):
-        package_advisories = PackageAdvisories()
-        package = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-
-        self.assertEqual(len(package_advisories), 0)
-
-        package_advisories.add_advisory_for_package(package, advisory)
-
-        self.assertEqual(len(package_advisories), 1)
-
-    def test_add_duplicate_advisory_for_package(self):
-        package_advisories = PackageAdvisories()
-        package1 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        package2 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory1 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-        advisory2 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-
-        self.assertEqual(len(package_advisories), 0)
-
-        package_advisories.add_advisory_for_package(package1, advisory1)
-
-        self.assertEqual(len(package_advisories), 1)
-
-        package_advisories.add_advisory_for_package(package1, advisory1)
-
-        self.assertEqual(len(package_advisories), 1)
-
-        package_advisories.add_advisory_for_package(package1, advisory2)
-
-        self.assertEqual(len(package_advisories), 1)
-
-        package_advisories.add_advisory_for_package(package2, advisory1)
-
-        self.assertEqual(len(package_advisories), 1)
-
-        package_advisories.add_advisory_for_package(package2, advisory2)
-
-    def test_get_package_advisories_for_package(self):
-        package_advisories = PackageAdvisories()
-        package1 = Package(
-            'foo', '1.2.3', '3', Architecture.AARCH64, 'foo-1.2.3-3.aarch64'
-        )
-        package2 = Package(
-            'bar', '1.2.3', '3', Architecture.AARCH64, 'bar-1.2.3-3.aarch64'
-        )
-        severity_date = datetime(year=2021, month=1, day=2, hour=11, minute=11)
-        creation_date = datetime(year=2021, month=3, day=21, hour=10, minute=0)
-        last_modification = datetime.now()
-        cvss_v3 = "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:L/I:L/A:L"
-        severity = Severity(
-            origin="foo",
-            date=severity_date,
-            cvss_v3=cvss_v3,
-        )
-        advisory1 = Advisory(
-            oid="1.2.3.4.5",
-            title="Foo Bar",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-        advisory2 = Advisory(
-            oid="1.2.3.4.6",
-            title="Bar Foo",
-            creation_date=creation_date,
-            last_modification=last_modification,
-            advisory_id="1234",
-            advisory_xref="http://foo/1234",
-            severity=severity,
-        )
-
-        self.assertEqual(len(package_advisories), 0)
-
-        package_advisories.add_advisory_for_package(package1, advisory1)
-        package_advisories.add_advisory_for_package(package1, advisory2)
-        package_advisories.add_advisory_for_package(package2, advisory2)
-
-        advisories1 = package_advisories.get_package_advisories_for_package(
-            package1
-        )
-        advisories2 = package_advisories.get_package_advisories_for_package(
-            package2
-        )
-        self.assertEqual(len(advisories1), 2)
-        self.assertEqual(len(advisories2), 1)
-
-
-class OperatingSystemAdvisoriesTestCase(TestCase):
-    def test_constructor(self):
-        os_advisories = OperatingSystemAdvisories()
-
-        self.assertIsInstance(os_advisories.advisories, dict)
-        self.assertEqual(len(os_advisories), 0)
-
-    def test_immutability(self):
-        os_advisories = OperatingSystemAdvisories()
-
-        with self.assertRaises(FrozenInstanceError):
-            os_advisories.advisories = dict()
-
-    def test_set_package_advisory(self):
-        os_advisories = OperatingSystemAdvisories()
-
-        self.assertEqual(len(os_advisories), 0)
-
-        advisories1 = PackageAdvisories()
-        os_advisories.set_package_advisories('BarOS 1.0', advisories1)
-
-        self.assertEqual(len(os_advisories), 1)
-
-        advisories2 = PackageAdvisories()
-        os_advisories.set_package_advisories('BarOS 1.0', advisories2)
-
-        self.assertEqual(len(os_advisories), 1)
-
-        advisories3 = PackageAdvisories()
-        os_advisories.set_package_advisories('FooOS 2.0', advisories3)
-        self.assertEqual(len(os_advisories), 2)
-
-    def test_get_package_advisory(self):
-        os_advisories = OperatingSystemAdvisories()
-
-        self.assertEqual(len(os_advisories), 0)
-
-        advisories = os_advisories.get_package_advisories('BarOS 1.0')
-        self.assertEqual(len(advisories), 0)
-
-        advisories1 = PackageAdvisories()
-        advisories2 = PackageAdvisories()
-        os_advisories.set_package_advisories('BarOS 1.0', advisories1)
-        os_advisories.set_package_advisories('FooOS 2.0', advisories2)
-
-        self.assertEqual(
-            os_advisories.get_package_advisories('BarOS 1.0'), advisories1
-        )
-        self.assertEqual(
-            os_advisories.get_package_advisories('FooOS 2.0'), advisories2
-        )
