+++ tests/models/test_vulnerability.py
@@ -19,7 +19,7 @@ from datetime import datetime
 from unittest import TestCase
 
 from notus.scanner.models.advisory import Advisory, Severity
-from notus.scanner.models.package import Architecture, Package
+from notus.scanner.models.package import Architecture, RPMPackage
 from notus.scanner.models.vulnerability import PackageVulnerability
 
 
@@ -45,19 +45,11 @@ class PackageVulnerabilityTestCase(TestCase):
             severity=severity,
         )
 
-        package = Package(
-            'foo-bar',
-            '1.2.3',
-            '4',
-            Architecture.X86_64,
+        package = RPMPackage.from_full_name(
             'foo-bar-1.2.3-4.x86_64',
         )
 
-        fixed_package = Package(
-            'foo-bar',
-            '1.2.4',
-            '4',
-            Architecture.X86_64,
+        fixed_package = RPMPackage.from_full_name(
             'foo-bar-1.2.4-4.x86_64',
         )
 
@@ -74,13 +66,6 @@ class PackageVulnerabilityTestCase(TestCase):
 
         advisory = package_vulnerability.advisory
         self.assertEqual(advisory.oid, '1.2.3.4.5')
-        self.assertEqual(advisory.title, 'Foo Bar')
-        self.assertEqual(advisory.creation_date, creation_date)
-        self.assertEqual(advisory.last_modification, last_modification)
-        self.assertEqual(advisory.advisory_id, '1234')
-        self.assertEqual(advisory.advisory_xref, 'http://foo/1234')
-        self.assertEqual(advisory.severity.origin, 'foo')
-        self.assertEqual(advisory.severity.date, severity_date)
 
         package = package_vulnerability.package
         self.assertEqual(package.name, 'foo-bar')
